# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lintel",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/work-items": {
      "post": {
        "tags": [
          "work-items"
        ],
        "summary": "Create Work Item",
        "operationId": "work-items_create_work_item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Work-Items Create Work Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-items"
        ],
        "summary": "List Work Items",
        "operationId": "work-items_list_work_items",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Work-Items List Work Items"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{work_item_id}": {
      "get": {
        "tags": [
          "work-items"
        ],
        "summary": "Get Work Item",
        "operationId": "work-items_get_work_item",
        "parameters": [
          {
            "name": "work_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Work-Items Get Work Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "work-items"
        ],
        "summary": "Update Work Item",
        "operationId": "work-items_update_work_item",
        "parameters": [
          {
            "name": "work_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Work-Items Update Work Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "work-items"
        ],
        "summary": "Remove Work Item",
        "operationId": "work-items_remove_work_item",
        "parameters": [
          {
            "name": "work_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/variables": {
      "post": {
        "tags": [
          "variables"
        ],
        "summary": "Create Variable",
        "operationId": "variables_create_variable",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVariableRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Variables Create Variable"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "variables"
        ],
        "summary": "List Variables",
        "operationId": "variables_list_variables",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "environment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Variables List Variables"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/variables/{variable_id}": {
      "get": {
        "tags": [
          "variables"
        ],
        "summary": "Get Variable",
        "operationId": "variables_get_variable",
        "parameters": [
          {
            "name": "variable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variable Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Variables Get Variable"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "variables"
        ],
        "summary": "Update Variable",
        "operationId": "variables_update_variable",
        "parameters": [
          {
            "name": "variable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variable Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Variables Update Variable"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "variables"
        ],
        "summary": "Delete Variable",
        "operationId": "variables_delete_variable",
        "parameters": [
          {
            "name": "variable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variable Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateVariableRequest": {
        "properties": {
          "variable_id": {
            "type": "string",
            "title": "Variable Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": ""
          },
          "environment_id": {
            "type": "string",
            "title": "Environment Id",
            "default": ""
          },
          "is_secret": {
            "type": "boolean",
            "title": "Is Secret",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "CreateVariableRequest"
      },
      "CreateWorkItemRequest": {
        "properties": {
          "work_item_id": {
            "type": "string",
            "title": "Work Item Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "work_type": {
            "$ref": "#/components/schemas/WorkItemType",
            "default": "task"
          },
          "status": {
            "$ref": "#/components/schemas/WorkItemStatus",
            "default": "open"
          },
          "assignee_agent_role": {
            "type": "string",
            "title": "Assignee Agent Role",
            "default": ""
          },
          "thread_ref_str": {
            "type": "string",
            "title": "Thread Ref Str",
            "default": ""
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name",
            "default": ""
          },
          "pr_url": {
            "type": "string",
            "title": "Pr Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title"
        ],
        "title": "CreateWorkItemRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateVariableRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "is_secret": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Secret"
          }
        },
        "type": "object",
        "title": "UpdateVariableRequest"
      },
      "UpdateWorkItemRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "work_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkItemType"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkItemStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee_agent_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Agent Role"
          },
          "thread_ref_str": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Ref Str"
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name"
          },
          "pr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Url"
          }
        },
        "type": "object",
        "title": "UpdateWorkItemRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkItemStatus": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "in_review",
          "approved",
          "merged",
          "closed",
          "failed"
        ],
        "title": "WorkItemStatus"
      },
      "WorkItemType": {
        "type": "string",
        "enum": [
          "feature",
          "bug",
          "refactor",
          "task"
        ],
        "title": "WorkItemType"
      },
...
    }
  }
}